fix(linux): properly cast rc to double

+++ src/unix/core.c
@@ -1869,14 +1869,14 @@ unsigned int uv_available_parallelism(void) {
 #ifdef __linux__
   cpu_set_t set;
   long rc;
-  int rc_with_cgroup;
+  double rc_with_cgroup;
 
   memset(&set, 0, sizeof(set));
 
   /* sysconf(_SC_NPROCESSORS_ONLN) in musl calls sched_getaffinity() but in
   * glibc it's... complicated... so for consistency try sched_getaffinity()
   * before falling back to sysconf(_SC_NPROCESSORS_ONLN).
-    */
+  */
   if (0 == sched_getaffinity(0, sizeof(set), &set))
     rc = CPU_COUNT(&set);
   else
@@ -1884,10 +1884,10 @@ unsigned int uv_available_parallelism(void) {
 
   uv__cpu_constraint c = uv__get_constrained_cpu();
   if (c.period_length > 0) {
-    rc_with_cgroup = (int)((double)(c.quota_per_period / c.period_length) * c.proportions);
+    rc_with_cgroup = (double)c.quota_per_period / c.period_length * c.proportions;
     
     if (rc_with_cgroup < rc)
-      rc = rc_with_cgroup;
+      rc = (long)rc_with_cgroup; /* Casting is safe since rc_with_cgroup < rc < LONG_MAX */
   }
 
   if (rc < 1)
